temporary status before reverting sys.comm

+++ tests/regression.sh
@@ -1,37 +1,29 @@
 #!/bin/bash
 
-function process() {
-	rm *.dat *.inp *.xyz *.rest 
-	
-	cp examples/${ref}.* .
-	
-	${exe} < ${ref}.inp
-	head -10 ${ref}.dat | awk '{ printf("%d %.6f %.6f %.6f\n", $1, $2, $3, $4); }' > a.dat
-	head -10 reference/${ref}.dat | awk '{ printf("%d %.6f %.6f %.6f\n", $1, $2, $3, $4); }' > b.dat
-	#cmp a.dat b.dat || exit 1
-	check=$(paste a.dat b.dat | awk 'function abs(x) {return (x<0?-x:x);} {for (i=1; i<5; ++i) {j=i+4; if (abs($i-$j) > '${tolerance}') {printf("failed %.6f vs %.6f\n", $i, $j);} } }')
-	
-	rm -f a.dat b.dat
-	
-	if [ -n "${check}" ]
-	then
-		echo "test on ${ref} failed"
-		exit 1
-	else
-		echo "test on ${ref} passed under tolerance ${tolerance}"
-	fi
-}
+tolerance=0.0001
 
-#exe=../ljmd.x
-exe=${1}
-tolerance=0.001
+rm *.dat *.inp *.xyz *.rest 
 
-#compare with references
-ref=argon_108
-process
+cp examples/${2}.* .
 
-ref=argon_2916
-process
+${1} < ${2}.inp
+res=$?
+if ((res != 0))
+then
+	exit 1
+fi
+head -10 ${2}.dat | awk '{ printf("%d %.6f %.6f %.6f\n", $1, $2, $3, $4); }' > a.dat
+head -10 reference/${2}.dat | awk '{ printf("%d %.6f %.6f %.6f\n", $1, $2, $3, $4); }' > b.dat
+check=$(paste a.dat b.dat | awk 'function abs(x) {return (x<0?-x:x);} {for (i=1; i<5; ++i) {j=i+4; if (abs($i-$j) > '${tolerance}') {printf("failed %.6f vs %.6f\n", $i, $j);} } }')
 
+rm -f a.dat b.dat
+
+if [ -n "${check}" ]
+then
+	echo "test on ${2} failed"
+	exit 1
+else
+	echo "test on ${2} passed under tolerance ${tolerance}"
+fi
 
 exit 0
